% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGnabit-package.R
\docType{package}
\name{DAGnabit-package}
\alias{DAGnabit}
\alias{DAGnabit-package}
\title{DAGnabit: Interactive DAG Visualization Widget}
\description{
Provides an HTML widget for creating interactive directed acyclic graphs (DAGs) in R.
}
\author{
\strong{Maintainer}: Tamara Babic \email{tbabic@dons.usfca.edu}

}
