% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagnabit.R
\name{dagnabit-shiny}
\alias{dagnabit-shiny}
\alias{dagnabitOutput}
\alias{renderDagnabit}
\title{Shiny bindings for DAGnabit}
\usage{
dagnabitOutput(outputId, width = "100\%", height = "600px")

renderDagnabit(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{Output variable to read from.}

\item{width, height}{Must be a valid CSS unit (like '100%' or '400px')
or a number that will be coerced to a string and have 'px' appended.}

\item{expr}{An expression that generates a DAGnabit widget.}

\item{env}{The environment in which to evaluate `expr`.}

\item{quoted}{Is `expr` a quoted expression (with `quote()`)? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using DAGnabit within Shiny
applications and interactive R Markdown documents.
}
