% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAGnabitWidget.R
\name{DAGnabitWidget}
\alias{DAGnabitWidget}
\title{DAGnabit HTML Widget}
\usage{
DAGnabitWidget(nodes, edges, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{nodes}{Data frame of nodes (must contain column `label`)}

\item{edges}{Data frame of edges (must contain columns `from`, `to`)}

\item{width}{Widget width}

\item{height}{Widget height}

\item{elementId}{optional HTML id}
}
\description{
DAGnabit HTML Widget
}
