% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1. hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{A character string containing your name.}
}
\value{
Prints a friendly message.
}
\description{
A simple test function for the DAGnabit package.
}
