% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layoutWidget.R
\name{layoutWidget}
\alias{layoutWidget}
\title{Layout Editor Widget}
\usage{
layoutWidget(
  nodes = list(),
  edges = list(),
  enableFileInputs = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{nodes}{list of node objects}

\item{edges}{list of edge objects}

\item{enableFileInputs}{logical}

\item{width}{widget width}

\item{height}{widget height}
}
\description{
Layout Editor Widget
}
