% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{visDependency}
\alias{visDependency}
\title{Add vis.js dependency}
\usage{
visDependency()
}
\value{
An htmltools::htmlDependency object
}
\description{
Internal helper to attach vis-network JavaScript to DAGnabitWidget.
}
\keyword{internal}
