% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_graph.R
\name{parse_graph}
\alias{parse_graph}
\title{Parse nodes and edges from text input}
\usage{
parse_graph(nodes_txt, edges_txt)
}
\arguments{
\item{nodes_txt}{Character string describing nodes, e.g. "A|none\\nB|none"}

\item{edges_txt}{Character string describing edges, e.g. "from,to\\nA,B\\nB,C"}
}
\value{
A list with elements `nodes` and `edges` as data frames
}
\description{
Parse nodes and edges from text input
}
